feat:fixed bug with admin analytics to see only undeleted posts as we…

+++ isa-OnlyBuns/src/main/java/rs/ac/uns/ftn/informatika/rest/dto/AnalyticsCountsDTO.java
@@ -2,16 +2,16 @@ package rs.ac.uns.ftn.informatika.rest.dto;
 
 public class AnalyticsCountsDTO {
 
-    public int weeklyComments;
-    public int monthlyComments;
-    public int yearlyComments;
+    public long weeklyComments;
+    public long monthlyComments;
+    public long yearlyComments;
 
-    public int weeklyPosts;
-    public int monthlyPosts;
-    public int yearlyPosts;
+    public long weeklyPosts;
+    public long monthlyPosts;
+    public long yearlyPosts;
 
-    public AnalyticsCountsDTO(int weeklyComments, int monthlyComments, int yearlyComments,
-                              int weeklyPosts, int monthlyPosts, int yearlyPosts) {
+    public AnalyticsCountsDTO(long weeklyComments, long monthlyComments, long yearlyComments,
+                              long weeklyPosts, long monthlyPosts, long yearlyPosts) {
         this.weeklyComments = weeklyComments;
         this.monthlyComments = monthlyComments;
         this.yearlyComments = yearlyComments;
